+++ examples/raw-nodejs/hmac-get-balance.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const { DEFAULT_BASE_URL, endpoint, hmacHeaders, jsonRequest, printResponse } = require('./payxcommerce');
+
+const publicKey = 'YOUR_PUBLIC_KEY';
+const secretKey = 'YOUR_SECRET_KEY';
+
+jsonRequest('GET', endpoint(DEFAULT_BASE_URL, '/balance'), hmacHeaders(publicKey, secretKey, '')).then(printResponse).catch((error) => {
+  console.error(error.message);
+  process.exit(1);
+});